Move public headers to iris subdirectory
