+++ src/cxx/cvxgen/cvxgen_ldp.c
@@ -1,4 +1,4 @@
-#include "cvxgen_ldp.h"
+#include "iris/cvxgen_ldp.h"
 #include "solver.h"
 
 Vars vars;
