+++ src/cxx/cvxgen/cvxgen_ldp.cpp
@@ -1,5 +1,5 @@
 #include <stdexcept>
-#include "cvxgen_ldp.h"
+#include "iris/cvxgen_ldp.h"
 
 using namespace Eigen;
