+++ src/cxx/iris/geometry.h
